+++ src/ts/content_scripts/twitch-injector.ts
@@ -1,145 +1,78 @@
-import { AuthorType } from '../../js/constants';
+import { isVod, parseMessageElement } from '../twitch-parser';
+import { nodeIsElement } from '../utils';
+import { getFrameInfoAsync } from '../../submodules/chat/src/ts/chat-utils';
 
 const liveChatSelector = '.chat-room .chat-scrollable-area__message-container';
 const vodChatSelector = '.video-chat .video-chat__message-list-wrapper ul';
 
-const badgesSelector = '.chat-badge';
-const displayNameSelector = '.chat-author__display-name';
-
-const liveChatLineSelector = '.chat-line__no-background';
-const liveMessageProperty = 'chat-line-message-body';
-const vodMessageSelector = '.video-chat__message';
-
-const vodTimestampSelector = '.vod-message__header p';
-const textFragmentClass = 'text-fragment';
-const emoteSelector = 'img.chat-line__message--emote';
-
-const isVod = window.location.pathname.includes('/videos/');
-let messageCounter = 0;
-
-function currentTime(): string {
-  const date = new Date();
-  return date.toLocaleTimeString('en-US', { hour: '2-digit', minute: '2-digit' });
-}
-
-function nodeIsElement(node: Node): node is Element {
-  return node.nodeType === Node.ELEMENT_NODE;
-}
-
-function elementIsAnchor(element: Element): element is HTMLAnchorElement {
-  return element.tagName === 'A';
-}
-
-function getVodMesssageBody(message: Element): Element | undefined {
-  const body = message.querySelector(vodMessageSelector);
-  return body?.children[1];
-}
-
-function getLiveMessageBody(message: Element): HTMLElement | undefined {
-  const chatLine = message.querySelector(liveChatLineSelector);
-  if (chatLine == null) return;
-  for (let i = 0; i < chatLine.children.length; i++) {
-    const child = chatLine.children[i] as HTMLElement;
-    if (child.dataset.testSelector === liveMessageProperty) {
-      return child;
-    }
-  }
-}
-
-function parseMessageFragment(fragment: Element): Ytc.ParsedRun | undefined {
-  if (fragment.classList.contains(textFragmentClass)) {
-    return {
-      type: 'text',
-      text: fragment.textContent ?? ''
-    };
-  } else if (elementIsAnchor(fragment)) {
-    return {
-      type: 'link',
-      text: fragment.textContent ?? fragment.href,
-      url: fragment.href
-    };
-  }
-
-  const emote = fragment.querySelector<HTMLImageElement>(emoteSelector);
-  if (emote == null) return;
-  return {
-    type: 'emoji',
-    src: emote.src,
-    alt: emote.alt
-  };
-}
-
-function parseTypes(message: Element): Ltl.AuthorType {
-  const badges = message.querySelectorAll<HTMLImageElement>(badgesSelector);
-  let types = 0;
-  Array.from(badges).forEach((badge) => {
-    // Doesn't work on other languages outside of English
-    if (badge.alt.toLowerCase() === 'moderator') types |= AuthorType.moderator;
-    else if (badge.alt.toLowerCase() === 'verified') types |= AuthorType.verified;
-    else if (badge.alt.toLowerCase() === 'broadcaster') types |= AuthorType.owner;
-    else if (badge.alt.toLowerCase().includes('subscriber')) types |= AuthorType.member;
-  });
-  return types;
-}
-
-function processMessages(message: Element): Ltl.Message | undefined {
-  const author = message.querySelector(displayNameSelector)?.textContent ?? '';
-  const timestamp = isVod ? message.querySelector(vodTimestampSelector)?.textContent ?? '' : currentTime();
-
-  const messageBody = isVod ? getVodMesssageBody(message) : getLiveMessageBody(message);
-  if (messageBody == null) return;
-  // console.debug({ messageBody });
-
-  const messageArray: Ytc.ParsedRun[] = [];
-  let text = '';
-  Array.from(messageBody.children).forEach((fragment) => {
-    // console.debug({ fragment });
-    const result = parseMessageFragment(fragment);
-    if (result == null) return;
-    if (result.type !== 'emoji') text += result.text;
-    messageArray.push(result);
-  });
-
-  const result = {
-    author,
-    timestamp,
-    messageArray,
-    authorId: author,
-    messageId: `${messageCounter++}`,
-    text,
-    types: parseTypes(message)
-  };
-  console.debug(result);
-  return result;
-}
-
-function chatMutationCallback(records: MutationRecord[]): void {
-  records.forEach((record) => {
-    const added = record.addedNodes;
-    if (added.length < 1) return;
-    added.forEach((node) => {
-      if (!nodeIsElement(node)) return;
-      processMessages(node);
-    });
-  });
-}
-
 let tries = 0;
+const clients: chrome.runtime.Port[] = [];
 
-function load(): void {
+function registerClient(port: chrome.runtime.Port): void {
+  if (clients.some((client) => client.name === port.name)) {
+    console.debug('Client already registered', { port, clients });
+    return;
+  }
+
+  port.onDisconnect.addListener(() => {
+    const i = clients.findIndex((clientPort) => clientPort.name === port.name);
+    if (i < 0) {
+      console.error('Failed to unregister client', { port, clients });
+      return;
+    }
+    clients.splice(i, 1);
+    console.debug('Unregister client successful', { port, clients });
+  });
+
+  clients.push(port);
+}
+
+async function load(): Promise<void> {
   const messageContainer = document?.querySelector(isVod ? vodChatSelector : liveChatSelector);
   // console.debug({ messageContainer });
   if (messageContainer == null) {
     if (tries++ < 5) {
-      setTimeout(load, 3000);
+      setTimeout(() => async () => await load(), 3000);
     } else {
       console.error('Could not find chat container');
     }
     return;
   }
 
-  const observer = new MutationObserver(chatMutationCallback);
+  const observer = new MutationObserver((mutationRecords) => {
+    mutationRecords.forEach((record) => {
+      const added = record.addedNodes;
+      if (added.length < 1) return;
+      added.forEach((node) => {
+        if (!nodeIsElement(node)) return;
+        const message = parseMessageElement(node);
+        if (message == null) return;
+        console.debug({ message });
+        clients.forEach((client) => client.postMessage({ type: 'message', message }));
+      });
+    });
+  });
+
+  chrome.runtime.onConnect.addListener((port) => {
+    port.onMessage.addListener((message) => {
+      switch (message.type) {
+        case 'registerClient':
+          registerClient(port);
+          break;
+        default:
+          console.error('Unknown message type', port, message);
+          break;
+      }
+    });
+  });
+
   observer.observe(messageContainer, { childList: true });
+
+  const frameInfo = await getFrameInfoAsync();
+
+  // TODO: pass frameInfo to popup/embed. Example:
+  // params.set('tabid', frameInfo.tabId.toString());
+  // params.set('frameid', frameInfo.frameId.toString());
 }
 
-load();
+load().catch((e) => console.error(e));
